FIX: Renamed auth
